+++ app/Services/Social/Meta/GraphPaginator.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Services\Social\Meta;
 
-use App\Exceptions\Social\IncompleteGraphPaginationException;
 use App\Services\Social\TokenRedactor;
 use Illuminate\Http\Client\ConnectionException;
 use Illuminate\Support\Facades\Http;
@@ -20,7 +19,7 @@ class GraphPaginator
      * @param  array<string, mixed>  $query
      * @return list<array<string, mixed>>
      *
-     * @throws IncompleteGraphPaginationException
+     * @throws IncompletePaginationException
      */
     public static function all(string $url, array $query = []): array
     {
@@ -79,12 +78,12 @@ class GraphPaginator
     /**
      * @return list<array<string, mixed>>
      *
-     * @throws IncompleteGraphPaginationException
+     * @throws IncompletePaginationException
      */
     private static function emptyOrIncomplete(int $successfulRequests, ?Throwable $previous = null): array
     {
         if ($successfulRequests > 0) {
-            throw new IncompleteGraphPaginationException($previous);
+            throw new IncompletePaginationException($previous);
         }
 
         return [];
