+++ app/Services/Social/Meta/IncompletePaginationException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Exceptions\Social;
+namespace App\Services\Social\Meta;
 
 use RuntimeException;
 use Throwable;
@@ -11,7 +11,7 @@ use Throwable;
  * Thrown when a Meta Graph edge was only partially fetched. Callers must not
  * treat a truncated list as complete (e.g. auto-connect when count === 1).
  */
-class IncompleteGraphPaginationException extends RuntimeException
+class IncompletePaginationException extends RuntimeException
 {
     public function __construct(?Throwable $previous = null)
     {
